+++ Application/Comms/DCommsStateUsbIdle.cpp
@@ -72,7 +72,7 @@ DCommsStateUsbIdle::DCommsStateUsbIdle(DDeviceSerial *commsMedium, DTask *task)
         PV624->handleError(E_ERROR_OS,
                            eSetError,
                            (uint32_t)os_error,
-                           (uint16_t)36);
+                           36u);
     }
 
     createCommands();
